refactor: move permission show/hide toggle in separate component
