+++ templates/Element/Form/trees.twig
@@ -18,14 +18,15 @@
                 {% set options = { 'multiple': 'true' }%}
             {% endif %}
 
+            {% set hasPermissions = 'Permissions' in schema.associations %}
+            <permission-toggle :user-roles="{{ user.roles|json_encode }}" :has-permissions="{{ hasPermissions|json_encode }}"></permission-toggle>
             <tree-view
                 relation-name={{ relationName }}
                 relation-label="{{ Layout.tr(relationName) }}"
                 :object='{{ { id: object.id, type: object.type }|json_encode }}'
                 :multiple-choice={{ options.multiple }}
                 :user-roles="{{ user.roles|json_encode }}"
-                :has-permissions="{{ ('Permissions' in schema.associations)|json_encode }}"
-                :show-forbidden="{{ _view.getRequest().getSession().read('showForbidden', true)|json_encode }}">
+                :has-permissions="{{ hasPermissions|json_encode }}">
             </tree-view>
             {% do Form.unlockField('relations.' ~ relationName ~ '.replaceRelated') %}
             {{ Form.hidden('_changedParents', {'value': '0', 'id': 'changedParents'})|raw }}
