+++ templates/Pages/Modules/index.twig
@@ -10,17 +10,8 @@
     <div class="module-index">
         {% if treeView %}
             {% set hasPermissions = 'Permissions' in schema.associations %}
-            {% if hasPermissions and not ('admin' in user.roles) %}
-            <div class="mb-05 p-1">
-                <span class="ml-05 mr-05"><: msgForbidden :></span>
-                <button class="button button-outlined" @click.prevent="toggleShowForbidden">
-                    <Icon icon="carbon:view" v-if="showForbidden"></Icon>
-                    <Icon icon="carbon:view-off" v-else></Icon>
-                    <Icon icon="carbon:locked" class="ml-05"></Icon>
-                </button>
-            </div>
-            {% endif %}
-            <tree-view :user-roles="{{ user.roles|json_encode }}" :has-permissions="{{ hasPermissions|json_encode }}" :show-forbidden="showForbidden"></tree-view>
+            <permission-toggle :user-roles="{{ user.roles|json_encode }}" :has-permissions="{{ hasPermissions|json_encode }}"></permission-toggle>
+            <tree-view :user-roles="{{ user.roles|json_encode }}" :has-permissions="{{ hasPermissions|json_encode }}"></tree-view>
         {% else %}
             <div class="table-container">
                 <div class="list-objects">
